+++ packages/ui/src/components/GridSettingsSpaceTab/GridSettingsSpaceTabStyles.ts
@@ -1,0 +1,11 @@
+import {createStyles, makeStyles, Theme} from "@material-ui/core/styles";
+
+const GridSettingsSpaceTabStyles = makeStyles((theme: Theme) =>
+  createStyles({
+    tabs: {
+      marginTop: theme.spacing(6),
+    },
+  }),
+);
+
+export default GridSettingsSpaceTabStyles;