+++ packages/ui/src/components/Profile/index.ts
@@ -1,2 +1,0 @@
-export {default} from "./Profile";
-export type {Props} from "./Profile";