+++ packages/ui/src/components/SpaceBilling/SpaceBilling.tsx
@@ -1,0 +1,39 @@
+import Error from "next/error";
+import * as React from "react";
+
+import {getSpace} from "@sentrei/common/firebase/spaces";
+import Space from "@sentrei/types/models/Space";
+import User from "@sentrei/types/models/User";
+import GridSettings from "@sentrei/ui/components/GridSettings";
+import SkeletonForm from "@sentrei/ui/components/SkeletonForm";
+
+export interface Props {
+  spaceId: string;
+  user: User.Get;
+}
+
+export default function SpaceSettings({spaceId}: Props): JSX.Element {
+  const [space, setSpace] = React.useState<Space.Get | null | undefined>();
+
+  React.useEffect(() => {
+    getSpace(spaceId).then(setSpace);
+  }, [spaceId]);
+
+  if (space === undefined) {
+    return (
+      <GridSettings skeleton tabSpaceKey="billing" type="space">
+        <SkeletonForm />
+      </GridSettings>
+    );
+  }
+
+  if (space === null) {
+    return <Error statusCode={404} />;
+  }
+
+  return (
+    <GridSettings spaceId={spaceId} tabSpaceKey="billing" type="space">
+      <></>
+    </GridSettings>
+  );
+}