+++ packages/ui/src/components/SpaceEdit/SpaceEdit.tsx
@@ -11,6 +11,7 @@ import Profile from "@sentrei/types/models/Profile";
 import Space from "@sentrei/types/models/Space";
 import User from "@sentrei/types/models/User";
 import FormSection from "@sentrei/ui/components/FormSection";
+import GridSettings from "@sentrei/ui/components/GridSettings";
 import SkeletonForm from "@sentrei/ui/components/SkeletonForm";
 import SpaceDescriptionForm from "@sentrei/ui/components/SpaceDescriptionForm";
 import SpaceNameForm from "@sentrei/ui/components/SpaceNameForm";
@@ -44,7 +45,7 @@ export default function SpaceEdit({
   }
 
   return (
-    <>
+    <GridSettings spaceId={spaceId} tabSpaceKey="general" type="space">
       <FormSection
         icon={<SettingsIcon />}
         title={t("space:space.editSpace")}
@@ -66,6 +67,6 @@ export default function SpaceEdit({
         }
         tabPanelThree={<></>}
       />
-    </>
+    </GridSettings>
   );
 }