+++ packages/web/pages_/[spaceId]/settings/quit.tsx
@@ -6,6 +6,7 @@ import * as React from "react";
 
 import AuthContext from "@sentrei/common/context/AuthContext";
 import {analytics} from "@sentrei/common/utils/firebase";
+import GridSettings from "@sentrei/ui/components/GridSettings";
 import SkeletonForm from "@sentrei/ui/components/SkeletonForm";
 import SentreiAppHeader from "@sentrei/web/components/SentreiAppHeader";
 
@@ -13,7 +14,7 @@ const SpaceQuit = dynamic(() => import("@sentrei/ui/components/SpaceQuit"), {
   ssr: false,
 });
 
-const SettingsQuitPage: NextPage = () => {
+const SpaceQuitPage: NextPage = () => {
   const {query} = useRouter();
 
   const {user, profile} = React.useContext(AuthContext);
@@ -26,7 +27,9 @@ const SettingsQuitPage: NextPage = () => {
     return (
       <>
         <SentreiAppHeader skeleton tabSpaceKey="settings" type="space" />
-        <SkeletonForm />
+        <GridSettings skeleton tabSpaceKey="quit" type="space">
+          <SkeletonForm />
+        </GridSettings>
       </>
     );
   }
@@ -52,4 +55,4 @@ const SettingsQuitPage: NextPage = () => {
   );
 };
 
-export default SettingsQuitPage;
+export default SpaceQuitPage;