+++ packages/web/pages_/profile.tsx
@@ -5,9 +5,9 @@ import * as React from "react";
 import AuthContext from "@sentrei/common/context/AuthContext";
 
 import {analytics} from "@sentrei/common/utils/firebase";
-import Loader from "@sentrei/ui/components/Loader";
 
 import ProfileEdit from "@sentrei/ui/components/ProfileEdit";
+import SkeletonForm from "@sentrei/ui/components/SkeletonForm";
 import SentreiAppHeader from "@sentrei/web/components/SentreiAppHeader";
 
 const Profile: NextPage = () => {
@@ -18,7 +18,12 @@ const Profile: NextPage = () => {
   }, []);
 
   if (user === undefined || !profile) {
-    return <Loader />;
+    return (
+      <>
+        <SentreiAppHeader skeleton tabUserKey="profile" type="user" />
+        <SkeletonForm />
+      </>
+    );
   }
 
   if (!user) {
@@ -32,6 +37,8 @@ const Profile: NextPage = () => {
           notificationCount={Number(user.notificationCount)}
           profile={profile}
           userId={user.uid}
+          tabUserKey="profile"
+          type="user"
         />
       )}
       {user && <ProfileEdit profile={profile} user={user} />}